+++ src/main/java/com/rogue/playtime/data/mysql/Data_MySQL.java
@@ -69,7 +69,7 @@ public class Data_MySQL implements DataHandler {
         return ret;
     }
 
-    public Map<String, Integer> getTopPlayers(String data, int amount) {
+    public Map<String, Integer> getTopPlayers(String data, byte amount) {
         this.db = new MySQL();
         Map<String, Integer> players = new HashMap();
         try {
