+++ src/main/java/com/rogue/playtime/runnable/EventRunnable.java
@@ -46,7 +46,7 @@ public class EventRunnable implements Runnable {
     private final String timer;
     private final List<String> commands;
     private final boolean repeat;
-    private static int counter = 0;
+    private static byte counter = 0;
 
     public EventRunnable(Playtime plugin, String eventName, String timer, Integer minimum, Integer maximum, List<String> commands, boolean repeat) {
         this.plugin = plugin;
